Added imem init files, and CPU testbench
Added the .mif file used for populating the IMEM(needs a restart of Xilinx ISE after each change), and the .coe file is used to initialize the memory when generating it. 
Also added the Top Module, CPU_Single_Cycle, which is essentially a functional MIPS(although with CHANIA Risk ISA) CPU.
NOTE THAT: branching doesnt work! Every time b, bne or beq is called, the Isim crashes.